# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
  /admin/auth/rotate:
    post:
      operationId: rotateAdminPassword
      tags: [admin]
      summary: Rotate admin password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotatePasswordBody"
      responses:
        "200":
          description: Password rotated, returns new token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get usage stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/logs:
    get:
      operationId: getAdminLogs
      tags: [admin]
      summary: Get recent request logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestLog"
  /admin/prompts:
    get:
      operationId: listAdminPrompts
      tags: [admin]
      summary: List practice prompts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticePrompt"
    post:
      operationId: createAdminPrompt
      tags: [admin]
      summary: Create a practice prompt
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticePromptBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticePrompt"
  /admin/prompts/{id}:
    put:
      operationId: updateAdminPrompt
      tags: [admin]
      summary: Update a practice prompt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticePromptBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticePrompt"
    delete:
      operationId: deleteAdminPrompt
      tags: [admin]
      summary: Delete a practice prompt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/grammar:
    get:
      operationId: listAdminGrammar
      tags: [admin]
      summary: List grammar topics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of grammar topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrammarTopic"
    post:
      operationId: createAdminGrammar
      tags: [admin]
      summary: Create a grammar topic
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrammarTopicBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrammarTopic"
  /admin/grammar/{id}:
    put:
      operationId: updateAdminGrammar
      tags: [admin]
      summary: Update a grammar topic
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrammarTopicBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrammarTopic"
    delete:
      operationId: deleteAdminGrammar
      tags: [admin]
      summary: Delete a grammar topic
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/security-events:
    get:
      operationId: listAdminSecurityEvents
      tags: [admin]
      summary: List security activity events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Security events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
  /admin/quiz/categories:
    get:
      operationId: listAdminQuizCategories
      tags: [admin]
      summary: List quiz categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of quiz categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizCategory"
    post:
      operationId: createAdminQuizCategory
      tags: [admin]
      summary: Create quiz category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizCategory"
  /admin/quiz/categories/{id}:
    put:
      operationId: updateAdminQuizCategory
      tags: [admin]
      summary: Update quiz category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizCategory"
    delete:
      operationId: deleteAdminQuizCategory
      tags: [admin]
      summary: Delete quiz category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/quiz/questions:
    get:
      operationId: listAdminQuizQuestions
      tags: [admin]
      summary: List quiz questions
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
    post:
      operationId: createAdminQuizQuestion
      tags: [admin]
      summary: Create quiz question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
  /admin/quiz/questions/{id}:
    put:
      operationId: updateAdminQuizQuestion
      tags: [admin]
      summary: Update quiz question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
    delete:
      operationId: deleteAdminQuizQuestion
      tags: [admin]
      summary: Delete quiz question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/exams:
    get:
      operationId: listAdminExams
      tags: [admin]
      summary: List exam editions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of exam editions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamEdition"
    post:
      operationId: createAdminExam
      tags: [admin]
      summary: Create exam edition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamEditionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamEdition"
  /admin/exams/{id}:
    put:
      operationId: updateAdminExam
      tags: [admin]
      summary: Update exam edition
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamEditionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamEdition"
    delete:
      operationId: deleteAdminExam
      tags: [admin]
      summary: Delete exam edition
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/wotd:
    get:
      operationId: listAdminWotd
      tags: [admin]
      summary: List word of the day entries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of WOTD entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WotdEntry"
    post:
      operationId: createAdminWotd
      tags: [admin]
      summary: Create WOTD entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WotdEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WotdEntry"
  /admin/wotd/{id}:
    put:
      operationId: updateAdminWotd
      tags: [admin]
      summary: Update WOTD entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WotdEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WotdEntry"
    delete:
      operationId: deleteAdminWotd
      tags: [admin]
      summary: Delete WOTD entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/config:
    get:
      operationId: getAdminConfig
      tags: [admin]
      summary: Get AI config
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Config values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConfig"
    put:
      operationId: updateAdminConfig
      tags: [admin]
      summary: Update AI config
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConfig"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    SecurityEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
      required:
        - id
        - type
        - description
        - timestamp
    RotatePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        ok:
          type: boolean
      required:
        - token
        - ok
...
    AdminStats:
      type: object
      properties:
        totalRequests:
          type: number
        aiCallsToday:
          type: number
        promptCallsToday:
          type: number
        wordOfDayCallsToday:
          type: number
        errorsToday:
          type: number
        uptimeSeconds:
          type: number
        startedAt:
          type: string
        requestsByEndpoint:
          type: object
          additionalProperties:
            type: number
        hourlyAiCalls:
          type: array
          items:
            $ref: "#/components/schemas/HourlyAiStat"
      required:
        - totalRequests
        - aiCallsToday
        - promptCallsToday
        - wordOfDayCallsToday
        - errorsToday
        - uptimeSeconds
        - startedAt
        - requestsByEndpoint
        - hourlyAiCalls
    RequestLog:
      type: object
      properties:
        id:
          type: string
        method:
          type: string
        path:
          type: string
        status:
          type: number
        durationMs:
          type: number
        timestamp:
          type: string
        isError:
          type: boolean
      required:
        - id
        - method
        - path
        - status
        - durationMs
        - timestamp
        - isError
    PracticePrompt:
      type: object
      properties:
        id:
          type: string
        taskType:
          type: string
        genre:
          type: string
        source:
          type: string
        prompt:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - taskType
        - genre
        - source
        - prompt
        - active
        - createdAt
    PracticePromptBody:
      type: object
      properties:
        taskType:
          type: string
        genre:
          type: string
        source:
          type: string
        prompt:
          type: string
        active:
          type: boolean
      required:
        - taskType
        - genre
        - source
        - prompt
        - active
    GrammarTopic:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        explanation:
          type: string
        examples:
          type: array
          items:
            type: string
        tips:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - explanation
        - examples
        - tips
        - active
        - createdAt
    GrammarTopicBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        explanation:
          type: string
        examples:
          type: array
          items:
            type: string
        tips:
          type: array
          items:
            type: string
        active:
          type: boolean
      required:
        - title
        - category
        - explanation
        - examples
        - tips
        - active
...